tests/requirements.txt
requests
numpy
pandas
matplotlib
PyYAML